# assets/openapi.yaml
openapi: 3.0.3
info:
  title: report-nexus
  version: "1.0"
  description: |
    Report service of the BLogic POS platform: renders sales reports as Excel or PDF and mails them through
    the License API. PDF pages are printed by a Gotenberg container.

    Every JSON response is the envelope `{data, message, isError, statusCode}`. Validation failures answer 400
    with `message: "Validation failed"` and `data` listing each failing field. JSON bodies are capped at 20 MB
    (`MAX_JSON_BODY`); larger bodies get 413. Field names are matched case-insensitively.
servers:
  - url: /
security: []
...
paths:
...
  /api/export-pdf:
    post:
      tags: [Export]
      summary: Print HTML to PDF
      operationId: exportPdf
      description: |
        Prints `htmlContent` with Chromium using screen styles and backgrounds. Content wider than A4 landscape
        (841.89pt) is zoomed out to fit. Resources on private, loopback or link-local addresses are not fetched.
        Download name: `fileName` reduced to letters, digits and spaces, plus `.pdf`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PDFRequest'
                - required: [htmlContent, fileName]
      responses:
        "200":
          $ref: '#/components/responses/Pdf'
        "400":
          $ref: '#/components/responses/Validation'
        "413":
          $ref: '#/components/responses/TooLarge'
        "500":
          description: Error generating PDF file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        "503":
          $ref: '#/components/responses/Busy'
...
components:
  responses:
...
    Pdf:
      description: The PDF.
      headers:
        Content-Disposition:
          description: 'attachment; filename=<name>.pdf (unquoted, may contain spaces)'
          schema:
            type: string
      content:
        application/pdf:
          schema:
            type: string
            format: binary
...
    Validation:
      description: Validation failed. `data` lists each failing field.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    TooLarge:
      description: Request body too large (20 MB).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    Busy:
      description: Server busy, please retry. The PDF renderer's queue is full, it timed out, or Chromium crashed.
      headers:
        Retry-After:
          schema:
            type: string
            example: "5"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
  schemas:
    Envelope:
      type: object
      properties:
        data:
          description: Payload; null on errors.
        message:
          type: string
        isError:
          type: boolean
        statusCode:
          type: integer
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  msg:
                    type: string
                    example: Meta From Date is required
                  path:
                    type: string
                    example: meta.fromDate
      example:
        data:
          - msg: Meta is required
            path: meta
        message: Validation failed
        isError: true
        statusCode: 400
...
    PDFRequest:
      type: object
      properties:
        htmlContent:
          type: string
          description: "A full HTML document, or the same document gzip-compressed and base64-encoded."
        fileName:
          type: string
          description: "Download name. A trailing .pdf is dropped and only letters, digits and spaces are kept."
        format:
          type: string
          enum: [letter, legal, tabloid, ledger, a0, a1, a2, a3, a4, a5, a6]
          description: "Paper name, case-insensitive. Wins over width/height. Default Letter (8.5 x 11 in)."
        width:
          oneOf:
            - type: number
            - type: string
          description: "Pixels when a number (96 per inch); a CSS length ending in px, in, cm or mm when a string. Both must be given to take effect; ignored when format is set."
        height:
          oneOf:
            - type: number
            - type: string
          description: "Pixels when a number (96 per inch); a CSS length ending in px, in, cm or mm when a string. Both must be given to take effect; ignored when format is set."
        landscape:
          type: boolean
          description: "Default false."
        margin:
          $ref: '#/components/schemas/Margin'
        datasets:
          $ref: '#/components/schemas/Datasets'
    Margin:
      type: object
      properties:
        top:
          oneOf:
            - type: number
            - type: string
          description: "Same units as width. Empty or missing means 0."
        right:
          oneOf:
            - type: number
            - type: string
          description: "Same units as width. Empty or missing means 0."
        bottom:
          oneOf:
            - type: number
            - type: string
          description: "Same units as width. Empty or missing means 0."
        left:
          oneOf:
            - type: number
            - type: string
          description: "Same units as width. Empty or missing means 0."
    Datasets:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/StoreMeta'
        tables:
          description: "Tables keyed by name, or a plain list. Rendered in this order."
          oneOf:
            - type: object
              additionalProperties:
                $ref: '#/components/schemas/Table'
            - type: array
              items:
                $ref: '#/components/schemas/Table'
    StoreMeta:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        dates:
          type: array
          items:
            type: string
          description: "Two ISO-8601 strings, the period start and end, printed under the title."
        date:
          type: string
          description: "Filled from dates before rendering; ignored on input."
    Table:
      type: object
      properties:
        title:
          type: string
          description: "Printed in upper case above the table."
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        datasets:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/Cell'
          description: "Rows; each row lists its cells in label order."
    Label:
      type: object
      properties:
        prop:
          type: string
        name:
          type: string
        type:
          type: string
          description: "string, number, currency, int, date or percent; drives the column width and alignment classes."
        bold:
          type: boolean
    Cell:
      type: object
      properties:
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
          description: "May be null. Numbers are printed as JavaScript would (1000000, not 1e+06)."
        isTextRight:
          type: boolean
        isBold:
          type: boolean
        link:
          type: string
          description: "When set the cell is rendered as a link to this URL."
...